pyproject.toml
[project]
name = "ml-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "hydra-core>=1.3.2",
    "rich>=14.3.3",
    "typer>=0.24.1",
    "wandb>=0.25.0",
]

[project.scripts]
ml-crucible = "cli:cli_app"
